Can create new items in cart, can show name and price.

+++ app/db/migrate/20190819092412_items.rb
@@ -3,10 +3,9 @@ class Items < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.string :name
       t.text :category
-      t.integer :price
+      t.decimal :price
       t.text :picture
       t.timestamps
     end
   end
 end
-
